Rename AnalyticsCollectionProxy to BulkAnalyticsCollectionProxy to accurately reflect the collection name
